+++ CHANGELOG.md
@@ -12,3 +12,4 @@ All noteable changes to this project will be documented in this file.
 - LICENSE files for hyper, alps, and constructor
 - Example for using the constructor library.
 - Collection+JSON consumer and producer architecture documents.
+- Collection+JSON producer implementation